requirements.txt
Django==1.7.1
Markdown==2.5.1
Pygments==1.6
Twisted==14.0.0
django-allauth==0.18.0
django-filter==0.8
django-localflavor==1.0
django-phonenumber-field==0.6
django-rest-auth==0.3.0
django-tokenapi==0.2.2
djangorestframework==2.4.3
oauthlib==0.7.1
phonenumbers==7.0.0
psutil==2.1.1
python-openid==2.2.5
requests==2.4.3
requests-oauthlib==0.4.2
wsgiref==0.1.2
zope.interface==4.1.1
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
python-gcm==0.1.5
